# Api-Backend-LocSystem/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "LocSystem API",
        "description": "Documentação da API do sistema LocSystem",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/my-deputies": {
            "get": {
                "tags": [
                    "Prepostos"
                ],
                "summary": "Lista os prepostos acessiveis ao usuario autenticado com filtros por busca e periodo.",
                "operationId": "a0aca4d8fa9aa0c7625d7c4581843c65",
                "responses": {
                    "200": {
                        "description": "Prepostos retornados com sucesso no campo deputies."
                    },
                    "401": {
                        "description": "Acesso negado. Apenas administradores e operadores podem consultar prepostos."
                    },
                    "500": {
                        "description": "Ocorreu um erro inesperado ao listar os prepostos."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Prepostos"
                ],
                "summary": "Cadastra um novo preposto vinculado ao operador autenticado.",
                "operationId": "e796469c955acd3a0136fddab9112d9b",
                "responses": {
                    "201": {
                        "description": "Preposto cadastrado com sucesso e retornado no campo deputy."
                    },
                    "401": {
                        "description": "Acesso negado. Apenas operadores podem cadastrar prepostos."
                    },
                    "409": {
                        "description": "Operador sem permissao operacional, e-mail ja utilizado ou limite de prepostos atingido."
                    },
                    "422": {
                        "description": "Erro de validacao nos dados enviados para cadastro do preposto."
                    },
                    "500": {
                        "description": "Ocorreu um erro inesperado ao cadastrar o preposto."
                    }
                },
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "v_name": {
                                        "type": "string",
                                        "minLength": 3,
                                        "maxLength": 255
                                    },
                                    "v_email": {
                                        "type": "string",
                                        "format": "email",
                                        "maxLength": 255
                                    },
                                    "v_phone": {
                                        "type": "string",
                                        "maxLength": 11
                                    },
                                    "v_password": {
                                        "type": "string",
                                        "minLength": 8,
                                        "maxLength": 16
                                    }
                                },
                                "required": [
                                    "v_name",
                                    "v_email",
                                    "v_phone",
                                    "v_password"
                                ]
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/my-deputies/{id}/approval": {
            "patch": {
                "tags": [
                    "Prepostos"
                ],
                "summary": "Aprova um preposto para acesso ao sistema.",
                "operationId": "9cc1bd09d965fd6419596cf49aec3850",
                "responses": {
                    "200": {
                        "description": "Preposto aprovado com sucesso e retornado no campo deputy."
                    },
                    "401": {
                        "description": "Acesso negado ou preposto nao encontrado."
                    },
                    "500": {
                        "description": "Ocorreu um erro inesperado ao aprovar o preposto."
                    }
                },
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "Parametro de rota id.",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/my-deputies/{id}/status": {
            "patch": {
                "tags": [
                    "Prepostos"
                ],
                "summary": "Bloqueia ou desbloqueia um preposto conforme o status informado.",
                "operationId": "03642cd6c91778a4354d81e925f8cc62",
                "responses": {
                    "200": {
                        "description": "Status do preposto atualizado com sucesso e dados retornados no campo deputy."
                    },
                    "401": {
                        "description": "Acesso negado ou preposto nao encontrado."
                    },
                    "409": {
                        "description": "O desbloqueio nao pode ser realizado antes da aprovacao, com operador bloqueado ou quando a assinatura nao estiver ativa."
                    },
                    "422": {
                        "description": "Erro de validacao no payload de alteracao de status do preposto."
                    },
                    "500": {
                        "description": "Ocorreu um erro inesperado ao atualizar o status do preposto."
                    }
                },
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "Parametro de rota id.",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "isActive": {
                                        "type": "boolean"
                                    },
                                    "reason": {
                                        "type": "string",
                                        "maxLength": 1000
                                    }
                                },
                                "required": [
                                    "isActive"
                                ]
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/my-deputies/{id}/reset-2fa": {
            "post": {
                "tags": [
                    "Prepostos"
                ],
                "summary": "Reseta a configuracao de autenticacao em dois fatores de um preposto.",
                "operationId": "6b0c62763843b74fa486e0e39ea8fe39",
                "responses": {
                    "200": {
                        "description": "2FA resetado com sucesso e confirmacao retornada no campo deputy."
                    },
                    "401": {
                        "description": "Acesso negado ou preposto nao encontrado."
                    },
                    "500": {
                        "description": "Ocorreu um erro inesperado ao resetar o 2FA do preposto."
                    }
                },
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "Parametro de rota id.",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}